Cloud Agent Space/requirements.txt
playwright==1.40.0
requests==2.31.0
flask==3.1.2
werkzeug==3.1.3
pandas==2.1.4
openpyxl==3.1.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

